requirements.txt
numpy==2.1.3
pygame-ce==2.5.2
